+++ package.json
@@ -37,7 +37,6 @@
     "font-awesome": "^4.7.0",
     "ionic-angular": "3.9.2",
     "ionicons": "3.0.0",
-    "jquery": "^3.3.1",
     "leaflet": "^1.3.4",
     "leaflet-search": "^2.9.7",
     "leaflet.locatecontrol": "^0.65.1",
