--- src/pages/stats/stats.html
@@ -1,30 +1,0 @@
-<!--
-  Generated template for the LeafletPage page.
-
-  See http://ionicframework.com/docs/components/#navigation for more info on
-  Ionic pages and navigation.
--->
-
-<!-- Import Leaflet CSS-->
-<link rel="stylesheet" href="https://unpkg.com/leaflet@1.3.4/dist/leaflet.css"
-integrity="sha512-puBpdR0798OZvTTbP4A8Ix/l+A4dHDD0DGqYW6RQ+9jxkRFclaxxQb/SJAWZfWAkuyeQUytO7+7N4QKrDh+drA=="
-crossorigin=""/>
-<!-- Import Leaflet.JS-->
-<script src="https://unpkg.com/leaflet@1.3.4/dist/leaflet.js"
-integrity="sha512-nMMmRyTVoLYqjP9hrbed9S+FzjZHW5gY1TWCHA5ckwXZBadntCNs8kEqAWdrb9O7rxbCaA4lKTIWjDXZxflOcA=="
-crossorigin=""></script>
-
-<ion-header>
-
-  <ion-navbar>
-    <ion-title>Graphs</ion-title>
-  </ion-navbar>
-
-</ion-header>
-
-
-<ion-content no-padding>
-
-  <canvas id="boxChart" width="400" height="400"></canvas>
-
-</ion-content>
