+++ src/providers/api/api.ts
@@ -6,7 +6,7 @@ export class ApiProvider {
 
   private API_URL = 'https://api.opensensemap.org/boxes';
   private exposure = '?exposure=outdoor';
-  private boxId = '/5a687677411a790019444133';
+  private boxId = '5a687677411a790019444133';
   private sensorId = '5a687677411a79001944413a';
   private sat24Location = 'DE'; // https://en.sat24.com/en/freeimages
   private sat24Url ='https://api.sat24.com/animated/'+this.sat24Location+'/rainTMC/2/Central%20European%20Standard%20Time/9553668';
